requirements.txt
# PDF Processing
# Replaced PyPDF2 with pypdf as recommended (PyPDF2 is deprecated)
pypdf>=3.15.1
pdfminer.six>=20220524
pdf2image>=1.16.0
pytesseract>=0.3.9

# Office Document Processing
python-docx>=0.8.11
openpyxl>=3.0.10
xlrd>=2.0.1
python-pptx>=0.6.21

# Data Handling
pandas>=1.5.0

# HTML/XML Processing
beautifulsoup4>=4.11.1
html2text>=2020.1.16
lxml>=4.9.1
requests>=2.28.1

# NLP Utilities
nltk>=3.7
spacy>=3.4.0
regex>=2022.3.15

# Configuration and CLI
pyyaml>=6.0
tqdm>=4.64.1
click>=8.1.3
